+++ outputs.tf
@@ -170,13 +170,13 @@ output "lan_secondary_mac_address" {
 output "vsocket_primary_reboot_status" {
   description = "Status of the Primary vSocket VM Reboot"
   value       = "Reboot triggered via Terraform"
-  depends_on  = [null_resource.reboot_vsocket_primary]
+  depends_on  = [terraform_data.reboot_vsocket_primary]
 }
 
 output "vsocket_secondary_reboot_status" {
   description = "Status of the Secondary vSocket VM Reboot"
   value       = "Reboot triggered via Terraform"
-  depends_on  = [null_resource.reboot_vsocket_secondary]
+  depends_on  = [terraform_data.reboot_vsocket_secondary]
 }
 
 output "cato_license_site" {